// Web/RMA-Web-Shell/projects/medicare/src/app/preauth-manager/models/preauth-claim-detail.ts
export class PreAuthClaimDetail
{
  claimId: number;
  claimReferenceNumber: string;
  claimLiabilityStatus: string;
  personEventId: number;
  eventId: number;
  isPensionCase: boolean;
  pensionCaseNumber: string;
  personName: string;
  dateOfBirth: Date;
  idNumber: string;
  passportNumber: string;
  employerName: string;
  industryNumber: string;
  claimContactNo: string;
  preAuthContactNo: string;
  eventTypeId: number;
  eventDate: Date;
  dateOfDeath: Date;
  isAlive: boolean;
  isStp: boolean;
}